Move client.ts into hot-reloader-client
